Create README.md
#Introduction
Quoting Wikipedia: "In biology, taxonomy is the science of naming, defining (circumscribing) and classifying groups of biological organisms on the basis of shared characteristics. Organisms are grouped together into taxa (singular: taxon) and these groups are given a taxonomic rank; groups of a given rank can be aggregated to form a super-group of higher rank, thus creating a taxonomic hierarchy. The principal ranks are domain, kingdom, phylum (division), class, order, family, genus, and species, but there are others as well allowing more fine-grained classification.

#Your task
In this assignment we will be working with an excerpt of the taxonomy of birds (i.e. the class of Aves) that are or have been encountered in New Zealand. For this purpose, we will access a data file that was downloaded from the New Zealand Organisms Register (NZOR) website, which can be found here. This data file contains information on different attributes of New Zealand birds, like their scientifc name, their - possibly several different - common names (vernacular names), their biostatus and their rank and classification in the above mentioned taxonomic hierarchy. For this assignment, you will be completing several code functions and classes using Coderunner3. These functions and classes will allow us to: represent birds using the information from the data file, read in this information from the file into our program, and manipulate/analyse the information.
The files containing the information about the bird taxonomy are ".csv" files, i.e. comma-separated-value files. These files are human readable text files (you can open them e.g. using Notepad++ or Excel), where information on one kind of bird is stored in a row, and different attributes (data fields) of a bird are stored in the columns, separated by a comma.